Move ActivityDetails to katpure-backend
