--- ui/src/components/activity/ActivityDetails/index.js
@@ -1,2 +1,0 @@
-import { ActivityDetails } from './connect';
-export { ActivityDetails };
