+++ ui/src/components/index.js
@@ -1,2 +1,4 @@
+export { Header } from './app/Header';
 export { ActivitySearch } from './activity/ActivitySearch';
 export { ActivityHeader } from './activity/ActivityHeader';
+export { PlateSizeForm } from './activity/PlateSizeForm';
