+++ ui/src/implementations/kapture-backend/components/ActivityDetails/ActivityDetails.module.css
