+++ ui/src/implementations/kapture-backend/components/ActivityDetails/Version.module.css
