+++ ui/src/implementations/kapture-backend/components/ActivityDetails/index.js
@@ -1,0 +1,1 @@
+export { ActivityDetails } from './connect';
