+++ ui/src/implementations/kapture-backend/components/index.js
@@ -1,3 +1,5 @@
-export { Header } from '../../../components/app/Header';
+export { Header } from '../../../components';
 export { ActivitySearch } from './ActivitySearchConnected';
 export { ActivityHeader } from './ActivityHeaderConnected';
+export { ActivityDetails } from './ActivityDetails';
+export { PlateSizeForm } from '../../../components';
